refactor: Improving your codebase with a modular structure
- Split a single codebase into seven specialised modules
 - utils.py: Common utility functions
 - filetree.py: File tree structure management
 - selector.py: File selection UI
 - output.py: Manage output formats
 - dependency.py: dependency analysis
 - cli.py: Command line interface
 - codeselect.py: Simplified entry point

- Added test code for each module
- Updated documentation (project_structure.md, change_log.md)
- Improved code readability and maintainability
- Clarified dependencies between modules

No functional changes, only structural improvements
