+++ docs/kr/task-log/refactor-module.md
@@ -16,11 +16,24 @@
 - ✅ **selector.py**: 파일 선택 UI (2025-03-10 완료)
   - `FileSelector` 클래스: curses 기반 대화형 파일 선택 UI
   - `interactive_selection()`: 선택 인터페이스 실행 함수
+- ✅ **output.py**: 출력 형식 관리 (2025-03-11 완료)
+  - `write_file_tree_to_string()`: 파일 트리를 문자열로 변환
+  - `write_output_file()`: 여러 형식의 출력 파일 생성 
+  - `write_markdown_output()`: 마크다운 형식 출력
+  - `write_llm_optimized_output()`: LLM 최적화 형식 출력
+- ✅ **dependency.py**: 의존성 분석 (2025-03-11 완료)
+  - `analyze_dependencies()`: 파일 간 의존성 분석
+- ✅ **cli.py**: 명령행 인터페이스 (2025-03-11 완료)
+  - `parse_arguments()`: 명령행 인수 처리
+  - `main()`: 메인 실행 함수
 
 ## 테스트 코드
 - ✅ **test/test_utils.py**: utils.py 기능 테스트 (2025-03-10 완료)
 - ✅ **test/test_filetree.py**: filetree.py 기능 테스트 (2025-03-10 완료)
 - ✅ **test/test_selector.py**: selector.py 기능 테스트 (2025-03-10 완료)
+- ✅ **test/test_output.py**: output.py 기능 테스트 (2025-03-11 완료)
+- ✅ **test/test_dependency.py**: dependency.py 기능 테스트 (2025-03-11 완료)
+- ✅ **test/test_cli.py**: cli.py 기능 테스트 (2025-03-11 완료)
 
 ## 남은 작업 및 파일 구조
 ```
@@ -29,9 +42,9 @@ codeselect/
 ├── utils.py          # 완료: 공통 유틸리티 함수 
 ├── filetree.py       # 완료: 파일 트리 구조 관리
 ├── selector.py       # 완료: 파일 선택 UI 
-├── output.py         # 예정: 출력 형식 관리
-├── dependency.py     # 예정: 의존성 분석
-└── cli.py            # 예정: 명령행 인터페이스
+├── output.py         # 완료: 출력 형식 관리
+├── dependency.py     # 완료: 의존성 분석
+└── cli.py            # 완료: 명령행 인터페이스
 ```
 
 ## 변환 작업 상세
@@ -53,18 +66,18 @@ codeselect/
    - `FileSelector` 클래스
    - `interactive_selection()` 함수
 
-4. **output.py**
+4. **output.py** ✅
    - `write_file_tree_to_string()` 함수
    - `write_output_file()` 함수
    - `write_markdown_output()` 함수
    - `write_llm_optimized_output()` 함수
 
-5. **dependency.py**
+5. **dependency.py** ✅
    - `analyze_dependencies()` 함수
 
-6. **cli.py**
+6. **cli.py** ✅
    - 명령행 인수 처리 (`argparse` 관련 코드)
    - `main()` 함수 리팩토링
 
-7. **codeselect.py** (리팩토링)
+7. **codeselect.py** ✅ (리팩토링)
    - 모듈들을 임포트하고 조합하는 간결한 메인 스크립트로 변환
